+++ source/game/blind_select.c
@@ -309,7 +309,7 @@ static void game_blind_select_display_blind_panel()
 
 static void game_blind_select_exit(void)
 {
-    reset_background();
+    change_background(BG_NONE, false);
     game_change_state(GAME_STATE_PLAYING);
 }
 
@@ -484,7 +484,7 @@ void game_blind_select_on_exit(void)
     sprite_destroy(&blind_select_tokens[BIG_BLIND]);
     sprite_destroy(&blind_select_tokens[BOSS_BLIND]);
 
-    reset_background();
+    change_background(BG_NONE, false);
     selection_y = 0;
 
     state_machine_remove(&blind_select_sm);
