+++ source/hand.c
@@ -9,6 +9,7 @@
 #include "audio_utils.h"
 #include "card.h"
 #include "game.h"
+#include "game/round.h"
 #include "game_variables.h"
 #include "graphic_utils.h"
 #include "soundbank.h"
@@ -124,6 +125,16 @@ enum HandType get_hand_type(void)
 ContainedHandTypes* get_contained_hands(void)
 {
     return &hand.contained_hands;
+}
+
+bool can_discard_hand(void)
+{
+    return (g_game_vars.discards > 0 && hand.state == HAND_SELECT && hand.hand_selections > 0);
+}
+
+bool can_play_hand(void)
+{
+    return (g_game_vars.hands > 0 && hand.state == HAND_SELECT && hand.hand_selections > 0);
 }
 
 static void print_hand_type(const char* hand_type_str)
