+++ source/joker_effects.c
@@ -1,4 +1,5 @@
 #include "game.h"
+#include "game/round.h"
 #include "game_variables.h"
 #include "hand.h"
 #include "joker.h"
